Moved Azure CLI file to Azure cluster provider directory
